App now owns lifetime of GameCommon objects; GameCommon objects are now strictly observing pointers; x64 works;

+++ Engine/Code/Engine/Core/Console.cpp
@@ -72,7 +72,7 @@ Console::Console(Renderer* renderer)
     select_all.cmd = IDM_SELECTALL;
 
     std::vector<ACCEL> accelerators = { copy, paste, cut, select_all };
-    hAcceleratorTable = ::CreateAcceleratorTableA(accelerators.data(), accelerators.size());
+    hAcceleratorTable = ::CreateAcceleratorTableA(accelerators.data(), static_cast<int>(accelerators.size()));
 }
 
 Console::~Console() {
